[planner] made the goal compiler more yarp friendly, some changes to planner
