+++ src/pradaPlanner-cpp/goalCompiler/CMakeLists.txt
@@ -16,18 +16,25 @@ find_package(ICUBcontrib)
 list(APPEND CMAKE_MODULE_PATH ${YARP_MODULE_PATH})
 list(APPEND CMAKE_MODULE_PATH ${ICUBCONTRIB_MODULE_PATH})
 
+# include(YarpIDL)
+# set(IDL_GENERATED_CODE_DIR "${CMAKE_CURRENT_SOURCE_DIR}") #?
+# yarp_idl_to_dir(${PROJECTNAME}.thrift ${IDL_GENERATED_CODE_DIR})
+
 include(YarpInstallationHelpers)
 include(ICUBcontribHelpers)
 include(ICUBcontribOptions)
 icubcontrib_set_default_prefix()
 
-file(GLOB source *.cpp)
+file(GLOB source src/*.cpp)
+file(GLOB header include/*.h)
 
 source_group("Source Files" FILES ${source})
+source_group("Header Files" FILES ${header})
 
-include_directories(${YARP_INCLUDE_DIRS})
+include_directories(${PROJECT_SOURCE_DIR}/include
+                    ${YARP_INCLUDE_DIRS})
 
-add_executable(${PROJECTNAME} ${source})
+add_executable(${PROJECTNAME} ${source} ${header})
 
 target_link_libraries(${PROJECTNAME} ${YARP_LIBRARIES})
